+++ catalyst/callbacks/metrics/auc.py
@@ -9,7 +9,7 @@ class AUCCallback(LoaderMetricCallback):
     """Calculates the AUC  per class for each loader.
 
     .. note::
-        Currently, supports binary and multi-label cases.
+        Currently, supports binary and multilabel cases.
     """
 
     def __init__(